refactor(systems/editor/buildings): Reorganizacion de carpetas
