+++ roguelike_project/systems/editor/buildings/docs/building_editor.md
